rename chatSchema to messageSchema, and response controller and model files
